requirements.txt
python-dotenv
openai
mistralai
pandas
numpy
streamlit
black